Move templates with paths to examples (#8643)
When a user creates a new document from a template, the template
is copied but relative paths are not changed, so the resulting
.lyx file is broken. By moving documents with relative paths to
examples, the files will compile out of the box.

A long-term solution that allows for relative paths in templates
is still desired and will be discussed in #8643 or in a new ticket.