AGU_article: Change defualt output format to xhtml
This document does not export to PDF. Further, if one tries
to export to PDF on some systems, the process hangs. This commit
thus changes the default output format to xhtml and explains in a
note at the top that LyX does not currently support exporting this
document to PDF.

Note also that this document seems to be outdated and so is a
candidate for attic.

Further, this document is a candidate for being renamed. Currently
users might be confused if they are looking for the canonical way
(no need for DocBook) to write an article for AGU journals. For
that, they should use AGUTeX.lyx.