Update RJournal.layout to v0.12 of RJournal.sty
Without this, the template gives a BibTeX error because
there are two \bibstyle commands in the .aux.

This would have broken backwards compatibility but RJournal.layout
has not yet been in a LyX release. Further, the journal made a
significant change (moving to a single column instead of two)
so any submitter would need to update the .sty anyway.