ctests for broken URLs in URL insets of LyX docs
These tests check for broken URLs in the URL insets of
the manuals, examples, and templates.

The tests are disabled by default because the Perl interpreter
is needed.

Later on they can be activated with a flag, as follows:

  cmake ... -DLYX_ENABLE_URLTESTS=ON

but for now the connection from the TOP-CMakeLists.txt is left out.

Missing part:
        1.) Declaring an setting the option
                LYX_OPTION(ENABLE_URLTESTS  "Enable for URL tests" OFF ALL)
        2.) make the connection
                if(LYX_ENABLE_URLTESTS)
                        add_subdirectory(development/checkurls "${TOP_BINARY_DIR}/checkurls")
                endif()