Add note to es/europeCV.lyx and invert Xe and Lua
Export with XeTeX and LuaTeX (with either non-tex fonts or 8-bit
compatibility mode) does not work because the loading of inputenc with
utf8x is hardcoded in europecv.cls at this time.

This commit adds a note to es/europeCV.lyx explaining the problem and
inverts the XeTeX and LuaTeX tests.

See
http://comments.gmane.org/gmane.editors.lyx.devel/145896

Thanks to Günter Milde for the advice and to Ignacio García
for the translation of the note.